+++ main-service/main-service-server/src/main/java/main/server/dto/location/LocationDto.java
@@ -16,9 +16,9 @@ public class LocationDto {
     @NotBlank
     String name;
     @NotNull
-    Float longitude;
+    Double longitude;
     @NotNull
-    Float latitude;
+    Double latitude;
     @NotNull
-    Float radius;
+    Double radius;
 }
